Generalize initSampleSetInsert for entity type
