+++ packages/components/src/components/samples/actions.ts
@@ -16,108 +16,13 @@
 import { Ajax, Filter, Utils } from '@labkey/api';
 import { fromJS, List, Map, OrderedMap } from 'immutable';
 
-import {
-    DisplayObject,
-    IParentOption,
-    IEntityTypeDetails,
-    IEntityTypeOption,
-    EntityIdCreationModel,
-    EntityTypeOption,
-    EntityParentType,
-} from '../entities/models';
-import { getSelected, getSelection } from '../../actions';
+import { IEntityTypeDetails, IParentOption, } from '../entities/models';
+import { getSelection } from '../../actions';
 import { SCHEMAS } from '../base/models/schemas';
-import { QueryColumn, SchemaQuery } from '../base/models/model';
+import { QueryColumn } from '../base/models/model';
 import { buildURL } from '../../url/ActionURL';
-import { getQueryGridModel } from '../../global';
 import { naturalSort } from '../../util/utils';
 import { selectRows } from '../../query/api';
-import { getActionErrorMessage } from "../../util/messaging";
-
-function initParents(initialParents: Array<string>, selectionKey: string): Promise<List<EntityParentType>> {
-    return new Promise((resolve) => {
-
-        if (selectionKey) {
-            const { schemaQuery } = SchemaQuery.parseSelectionKey(selectionKey);
-            const queryGridModel = getQueryGridModel(selectionKey);
-
-            if (queryGridModel && queryGridModel.selectedLoaded) {
-                return selectRows({
-                            schemaName: schemaQuery.schemaName,
-                            queryName: schemaQuery.queryName,
-                            columns: 'LSID,Name,RowId',
-                            filterArray: [Filter.create('RowId', queryGridModel.selectedIds.toArray(), Filter.Types.IN)]
-                        }).then((samplesResponse) => {
-                            resolve(resolveSampleSetParentTypeFromIds(schemaQuery, samplesResponse));
-                        });
-            }
-            else {
-                getSelected(selectionKey).then((selectionResponse) => {
-                    return selectRows({
-                        schemaName: schemaQuery.schemaName,
-                        queryName: schemaQuery.queryName,
-                        columns: 'LSID,Name,RowId',
-                        filterArray: [Filter.create('RowId', selectionResponse.selected, Filter.Types.IN)]
-                    }).then((samplesResponse) => {
-                        resolve(resolveSampleSetParentTypeFromIds(schemaQuery, samplesResponse));
-                    });
-                }).catch(() => {
-                    console.warn('Unable to parse selectionKey', selectionKey);
-                    resolve(List<EntityParentType>());
-                });
-            }
-        }
-        else if (initialParents && initialParents.length > 0) {
-            const parent = initialParents[0];
-            const [ schema, query, value ] = parent.toLowerCase().split(':');
-
-            return selectRows({
-                schemaName: schema,
-                queryName: query,
-                columns: 'LSID,Name,RowId',
-                filterArray: [Filter.create('RowId', value)]
-            }).then((samplesResponse) => {
-                resolve(resolveSampleSetParentTypeFromIds(SchemaQuery.create(schema, query), samplesResponse));
-            });
-        }
-        else {
-            resolve(List<EntityParentType>());
-        }
-    });
-}
-
-function resolveSampleSetParentTypeFromIds(schemaQuery: SchemaQuery, response: any): List<EntityParentType> {
-    const { key, models, orderedModels } = response;
-    const rows = fromJS(models[key]);
-    let data = List<DisplayObject>();
-
-    // The transformation done here makes the samples compatible with the editable grid
-    orderedModels[key].forEach((id) => {
-        const row = extractFromRow(rows.get(id));
-        data = data.push({
-            displayValue: row.label,
-            value: row.rowId
-        });
-    });
-
-    return List<EntityParentType>([
-        EntityParentType.create({
-            index: 1,
-            schema: schemaQuery.getSchema(),
-            query: schemaQuery.getQuery(),
-            value: data
-        })
-    ]);
-}
-
-function extractFromRow(row: Map<string, any>): IEntityTypeOption {
-    return {
-        label: row.getIn(['Name', 'value']),
-        lsid: row.getIn(['LSID', 'value']),
-        rowId: row.getIn(['RowId', 'value']),
-        value: row.getIn(['Name', 'value'])
-    }
-}
 
 export function initSampleSetSelects(isUpdate: boolean, ssName: string, placeholderOption: IParentOption, prefix: string, ) :Promise<List<IParentOption>> {
     return selectRows({
@@ -144,61 +49,6 @@ export function initSampleSetSelects(isUpdate: boolean, ssName: string, placehol
         }
 
         return sets.sortBy(p => p.label, naturalSort) as List<IParentOption>;
-    });
-}
-
-
-
-export function initSampleSetInsert(model: EntityIdCreationModel)  : Promise<Partial<EntityIdCreationModel>> {
-    return new Promise( (resolve, reject) => {
-        return Promise.all([
-            selectRows({
-                schemaName: SCHEMAS.EXP_TABLES.SAMPLE_SETS.schemaName,
-                queryName: SCHEMAS.EXP_TABLES.SAMPLE_SETS.queryName,
-                columns: 'LSID,Name,RowId'
-            }),
-            initParents(model.parents, model.selectionKey)
-        ]).then(results => {
-            let [ sampleSetResult, sampleParents ] = results;
-            const sampleSets = fromJS(sampleSetResult.models[sampleSetResult.key]);
-            const sampleCount = sampleParents.find((parent) => parent.value !== undefined) ? 1 : 0;
-            const parentOptions = sampleSets.map(row => {
-                const name = row.getIn(['Name', 'value']);
-                return {
-                    value: name.toLowerCase(),
-                    label: name,
-                    schema: SCHEMAS.SAMPLE_SETS.SCHEMA,
-                    query: name // Issue 33653: query name is case-sensitive for some data inputs (sample parents)
-                }
-            }).toList().sortBy(p => p.label, naturalSort);
-
-            const sampleSetOptions = sampleSets
-                .map(row => extractFromRow(row))
-                .sortBy(r => r.label, naturalSort)
-                .toList();
-
-            let targetSampleSet;
-            if (model.initialEntityType) {
-                const setName = model.initialEntityType.toLowerCase();
-                const data = sampleSets.find(row => setName === row.getIn(['Name', 'value']).toLowerCase());
-
-                if (data) {
-                    targetSampleSet = new EntityTypeOption(extractFromRow(data));
-                }
-            }
-            resolve({
-                isInit: true,
-                parentOptions,
-                entityCount: sampleCount,
-                entityParents: sampleParents,
-                entityTypeOptions: sampleSetOptions,
-                targetEntityType: targetSampleSet
-            })
-        })
-        .catch((reason) => {
-            console.error(reason);
-            reject(getActionErrorMessage('There was a problem initializing the sample type create page.', 'sample types'));
-        })
     });
 }
